refactor(api): replace `level` endpoint with `PATCH /api/priority/`
